+++ packages/client/vite.config.ts
@@ -22,7 +22,7 @@ export default defineConfig({
       fileName: format => `index.${format}.js`,
     },
     rollupOptions: {
-      external: [],
+      external: ["@graphand/core"],
       output: {
         globals: {},
       },